requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3
Flask-Cors==4.0.0
Flask-Swagger-UI==4.11.1
python-dotenv==1.0.0
pytest==7.4.0
pytest-flask==1.2.0
gunicorn==21.2.0